Initial commit
Basic queries work but the work to support transactions is not complete
yet and the work to support query parameters is not started. May try to
combine the service end points so that there is just one, rather than
four (or five, when the parameter support is completed).

This doesn't have any support for recording and playing back queries at
this point, it only proxies calls to a SQL instance.